refactor: organize shared components
