+++ src/InventoryBatchPreview.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState } from "react";
 import InventoryGrid from "./InventoryGrid";
-import { ViewToggle } from "./ViewToggle";
+import { ViewToggle } from "./shared/ViewToggle";
 import type { ViewMode } from "./types/ui";
 import "./InventoryBatchPreview.css";
